tweak sample logic and do some refactoring
